genpkglist.cc: rewrite copyStrippedFileList() to avoid inplace dirnames edit bug
There was an assumption that DIRINDEXES are always sorted ascending,
which actually does not hold.  This means we CANNOT use inplace
replacement in "dirnames" array, so as to append "later dirs" on top
of "older dirs".

The bug can actually be more subtle.  It is very well possible that
I simply cannot understand that previous "inplace optimization" logic.
But I think that it sucks anyway.  I choose to rewrite this piece of code
for the sake of clarity and correctness.  I hope that the correctness of
the code now should be a bit more obvious, even for a casual reader.