rpmpm.c: fixed rpmdepCheck() call
It is not clear what rpmdepCheck() return code is supposed to mean.
Logically, non-zero return code should mean that the dependency check
failed.  However, rpm can also return 0 (for "no error") while setting
"conflicts" pointer.  So, for now, it seems that we should check both
the return code and the pointer.