disabled alternative uncompressed URIs
When a compressed file: URI is requested, it checks to see if the
uncompressed file exsits, and uses that instead.  This isn't always
desireable, since e.g. zstd(1) is known for its very fast uncompression
which can be performed even as you read a file, but uncompressing the
whole file can eat up too much virtual memory.