pkgAcqIndex: support .zst and recompression to .lz4 (1)
This is the first patch of the series which adds support for zstd
compression.  By default, apt will select pkglist.classic$zext
compression method based on the Acquire::ComprExtension configuration
variable, which is ".xz" by default.  When a list.xz is downloaded,
it will be recompressed to .lz4 (rather than just uncompressed).
When Acquire::ComprExtension is set to ".zst", apt will look for
list.zst, but will still fall back to list.xz if list.zst is missing.
When a list.zst is downloaded, it will not be recompressed to .lz4,
because zstd decompression is fast on its own.

This patch is not self-sufficient, it only alters the pkgAcqIndex
constructor.  It imposes a slightly stricter requirement on a repo:
if a repo has a release file, then the release file must contain
all the necessary checksums (in other words, both pkglist.classic
and pkglist.classic$zext must be present in the release file).
It also adds an optimization: when list.lz4 or list.zst is up
to date, there is nothing to be done at all.